+++ src/test/routes/healthcheckRouter.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest'
-import app from '../app'
+import app from '../../app'
 
 describe('healthcheckController', () => {
   describe('GET /ping', () => {
